Run tests with J2Cl, by forking a Maven process
The tests are almost copies from the GWT tests, slightly
modified to account for asynchronous testing differences
and the lack of an embedded web server.
The servlets are thus run in a Jetty container with an
additional CORS filter to allow calls from the test pages
loaded as file:// URLs by the j2cl:test.